Create Type, Example and Definition classes and move its own methods
